RAG_QA_Finance_ML/requirements.txt
langchain
streamlit
pypdf
python-dotenv
langchain_community
langchain-pinecone
langchain-groq
langchain-core
sentence-transformers
pinecone-client
